[WIP] Extend idempotence test to actually send a payment handle request.
Allow to configure the PAYONE API URL in the ServiceConfig.

Use a WireMock proxy server in the idempotence test fixture to verify
that no request is sent to PAYONE in case of payments that are already
in TransactionState "Success" or "Failure".

Use Apache's fluent http client to send requests to the integration
service's HTTP API to force payment handling.