Add executor for unsupported transactions.
UnsupportedTransactionExecutor executes a transaction by setting the payment's transaction state to Failure
and adding an interfaceInteraction containing a reference to the transaction as well as a message
explaining the reason why it failed.

Some refactorings regarding dependency injection.

UnsupportedTransactionExecutorTest added but ignored due to our X-mas party taking place now :)