Refactor specification test fixtures to use service instance deployed to Heroku.
The tasks :functionaltests:test[Spec] require an environment variable or system property 'CT_PAYONE_INTEGRATION_URL' with the base
URL of the integration service instance to talk to.

The service's CustomTypeBuilder may be configured to erase all Order, Cart, Payment and Type instances so that it can create the custom types
from scratch. The environment variable or system property must be called 'CT_START_FROM_SCRATCH' and have a value of 'true' to grant the
permission to erase. In case of 'false' or an absent variable/property the permission won't be granted.

The test fixtures AuthorizationFixture and ChargePreauthorizedFixture have been refactored to use the deployed service instead of an
instance created by them.