Process "capture" transaction status notifications
CaptureNotificationProcessor adds a "PAYONE_INTERACTION_NOTIFICATION" to the respective payment for each "capture" transaction status notification.
If the payment does not contain a "Charge" transaction whose "interactionId" matches the notification's "sequencenumber" yet it adds a new one with state "Pending".
The transaction state of an already existing "Charge” transaction is not changed (assumed to be "Pending").