Fetch amountPaid in specification test for Charge w/o separate Authorization
The ChargeImmediatelyFixture didn't fetch the amountPaid but expected it to have a certain value.
BaseFixture's constant PAYONE_NOTIFICATION_TIMEOUT had to be increased to 8 minutes because during debugging I noticed a paid notification that was sent more than 6 minutes after the appointed notification.
To reduce the time the specification tests need to wait, AuthorizationFixture and ChargeImmediatelyFixture wait for a batch of payment notifications instead of per payment.
The ChargeImmediatelyFixture still fails because the amountAuthorized is not set correctly.

Contributes to #7.