[specs] Remove executable specification stubs for failed verifications because they're out of scope
The Integration Service is notified by PAYONE only in case of successful verification.
In case of failed/cancelled verifications the user is redirected to the URL in the payment's custom field errorUrl or backUrl (not in case of 3-D Secure).
The checkout implementation needs to take care of that.

Contributes to #23.