Use DefaultNotificationProcessor to simply persist notifications and report "TSOK" to PAYONE
The DefaultNotificationProcessor is invoked by the NotificationDispatcher if there is no specific NotificationProcesser for the particular "txaction".
It persists the notification by adding an interface interaction to the payment - no matter which "txaction" value was provided.
This is necessary to allow processing of transaction not (yet) supported by the integration service and to prevent the service from blocking the further processing of the payment (by replying "TSOK" to PAYONE's transaction status notification POST request).